+++ tests/component/session/conftest.py
@@ -235,7 +235,7 @@ async def session_seed(
     )
 
     async with db_engine.begin() as conn:
-        await conn.execute(kernels.delete().where(kernels.c.id == kernel_id))
+        await conn.execute(kernels.delete().where(kernels.c.session_id == session_id))
         await conn.execute(
             SessionRow.__table__.delete().where(SessionRow.__table__.c.id == session_id)
         )
@@ -330,7 +330,94 @@ async def terminated_session_seed(
     )
 
     async with db_engine.begin() as conn:
-        await conn.execute(kernels.delete().where(kernels.c.id == kernel_id))
+        await conn.execute(kernels.delete().where(kernels.c.session_id == session_id))
+        await conn.execute(
+            SessionRow.__table__.delete().where(SessionRow.__table__.c.id == session_id)
+        )
+
+
+@pytest.fixture()
+async def user_session_seed(
+    db_engine: SAEngine,
+    domain_fixture: str,
+    group_fixture: uuid.UUID,
+    regular_user_fixture: UserFixtureData,
+    scaling_group_fixture: str,
+) -> AsyncIterator[SessionSeedData]:
+    """Seed a RUNNING session owned by the regular user."""
+    unique = secrets.token_hex(4)
+    session_id = SessionId(uuid.uuid4())
+    session_name = f"test-user-session-{unique}"
+    kernel_id = uuid.uuid4()
+    now = datetime.now(tzutc())
+
+    status_history: dict[str, Any] = {
+        SessionStatus.PENDING.name: now.isoformat(),
+        SessionStatus.RUNNING.name: now.isoformat(),
+    }
+
+    async with db_engine.begin() as conn:
+        await conn.execute(
+            sa.insert(SessionRow.__table__).values(
+                id=session_id,
+                creation_id=f"cid-{unique}",
+                name=session_name,
+                session_type=SessionTypes.INTERACTIVE,
+                cluster_size=1,
+                cluster_mode="single-node",
+                domain_name=domain_fixture,
+                group_id=group_fixture,
+                user_uuid=regular_user_fixture.user_uuid,
+                access_key=regular_user_fixture.keypair.access_key,
+                scaling_group_name=scaling_group_fixture,
+                status=SessionStatus.RUNNING,
+                status_info="",
+                status_history=status_history,
+                occupying_slots=ResourceSlot(),
+                requested_slots=ResourceSlot(),
+                created_at=now,
+            )
+        )
+        await conn.execute(
+            sa.insert(kernels).values(
+                id=kernel_id,
+                session_id=session_id,
+                session_creation_id=f"cid-{unique}",
+                session_name=session_name,
+                session_type=SessionTypes.INTERACTIVE,
+                cluster_role="main",
+                cluster_idx=0,
+                cluster_hostname="main0",
+                cluster_mode="single-node",
+                cluster_size=1,
+                domain_name=domain_fixture,
+                group_id=group_fixture,
+                user_uuid=regular_user_fixture.user_uuid,
+                access_key=regular_user_fixture.keypair.access_key,
+                scaling_group=scaling_group_fixture,
+                status=KernelStatus.RUNNING,
+                status_info="",
+                occupied_slots=ResourceSlot(),
+                requested_slots=ResourceSlot(),
+                repl_in_port=0,
+                repl_out_port=0,
+                stdin_port=0,
+                stdout_port=0,
+                created_at=now,
+            )
+        )
+
+    yield SessionSeedData(
+        session_id=session_id,
+        session_name=session_name,
+        kernel_id=kernel_id,
+        access_key=regular_user_fixture.keypair.access_key,
+        domain_name=domain_fixture,
+        user_uuid=regular_user_fixture.user_uuid,
+    )
+
+    async with db_engine.begin() as conn:
+        await conn.execute(kernels.delete().where(kernels.c.session_id == session_id))
         await conn.execute(
             SessionRow.__table__.delete().where(SessionRow.__table__.c.id == session_id)
         )
